requirements.txt
flask==2.3.3
python-telegram-bot==20.8
python-dotenv==1.0.0
PyPDF2==3.0.1
gunicorn==21.2.0